backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1 

SQLAlchemy==2.0.19

python-dotenv==1.0.0
